Organise all classes into neat packages
