+++ docs/README.md
@@ -1,4 +1,4 @@
-# KleeBot User Guide
+# kleebot.kleebot User Guide
 
 // Update the title above to match the actual product name
