--- src/main/java/Task.java
@@ -1,105 +1,0 @@
-public abstract class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return isDone ? "X" : " ";
-    }
-
-    public boolean getStatus() { return isDone? true : false; }
-
-    public String getDescription() {
-        return this.description;
-    }
-    
-    public abstract String getType();
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void unmarkAsDone() {
-        this.isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + getStatusIcon() + "]" + getDescription();
-    }
-}
-
-
-
-
-
-
-
-
-class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    public String getType() {
-        return "T";
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
-
-class Deadline extends Task {
-    protected String by;
-
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = by.trim();
-    }
-
-    public String getType() {
-        return "D";
-    }
-
-    public String getBy() {
-        return this.by;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by + ")";
-    }
-}
-
-class Event extends Task {
-    protected String from;
-    protected String to;
-
-    public Event(String description, String from, String to) {
-        super(description);
-        this.from = from.trim();
-        this.to= to.trim();
-    }
-
-    public String getType() {
-        return "E";
-    }
-
-    public String getFrom() {
-        return from;
-    }
-    public String getTo() {
-        return to;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
-    }
-}
