+++ src/main/java/kleebot/command/EventCommand.java
@@ -1,0 +1,24 @@
+package kleebot.command;
+
+import kleebot.storage.Storage;
+import kleebot.task.Event;
+import kleebot.task.TaskList;
+import kleebot.ui.Ui;
+
+public class EventCommand extends Command {
+    private final String description;
+    private final String from;
+    private final String to;
+
+    public EventCommand(String description, String from, String to) {
+        this.description = description;
+        this.from = from;
+        this.to = to;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        Event event = new Event(description, from, to);
+        tasks.addToList(event);
+    }
+}
