+++ src/main/java/kleebot/command/ExitCommand.java
@@ -1,0 +1,17 @@
+package kleebot.command;
+
+import kleebot.storage.Storage;
+import kleebot.task.TaskList;
+import kleebot.ui.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.exit();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
