+++ src/main/java/kleebot/command/UnknownCommand.java
@@ -1,0 +1,13 @@
+package kleebot.command;
+
+import kleebot.storage.Storage;
+import kleebot.task.TaskList;
+import kleebot.ui.Ui;
+
+public class UnknownCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.throwTantrum();
+    }
+}
