+++ src/main/java/kleebot/task/Deadline.java
@@ -1,0 +1,23 @@
+package kleebot.task;
+
+public class Deadline extends Task {
+    protected String by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = by.trim();
+    }
+
+    public String getType() {
+        return "D";
+    }
+
+    public String getBy() {
+        return this.by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
