+++ src/main/java/kleebot/task/Event.java
@@ -1,0 +1,29 @@
+package kleebot.task;
+
+public class Event extends Task {
+    protected String from;
+    protected String to;
+
+    public Event(String description, String from, String to) {
+        super(description);
+        this.from = from.trim();
+        this.to = to.trim();
+    }
+
+    public String getType() {
+        return "E";
+    }
+
+    public String getFrom() {
+        return from;
+    }
+
+    public String getTo() {
+        return to;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
+    }
+}
