+++ src/main/java/kleebot/task/Task.java
@@ -1,0 +1,41 @@
+package kleebot.task;
+
+public abstract class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return isDone ? "X" : " ";
+    }
+
+    public boolean getStatus() { return isDone? true : false; }
+
+    public String getDescription() {
+        return this.description;
+    }
+    
+    public abstract String getType();
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void unmarkAsDone() {
+        this.isDone = false;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + getStatusIcon() + "]" + getDescription();
+    }
+}
+
+
+
+
+
