+++ src/main/java/kleebot/task/ToDo.java
@@ -1,0 +1,16 @@
+package kleebot.task;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    public String getType() {
+        return "T";
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
