+++ src/main/java/kleebot/ui/KleeExceptions.java
@@ -1,3 +1,5 @@
+package kleebot.ui;
+
 public class KleeExceptions extends Exception {
     public KleeExceptions(String m) {
         super(m);
