+++ src/main/java/kleebot/ui/Ui.java
@@ -1,3 +1,5 @@
+package kleebot.ui;
+
 public class Ui {
 
     public static final String LINE_BREAK = "____________________________________________________________";
@@ -19,7 +21,7 @@ public class Ui {
                     "⠀⠀⠀⠀⠈⢛⣿⣿⣿⡇⠀⠀⠀⠉⢻⣶⡟⠹⣷⣎⠉⠉⠉⠉⢻⠄⠀⠀⠀";
 
 
-    enum ErrorMessage {
+    public enum ErrorMessage {
         MISSING_DETAILS("Hey! Gimmie more details about this task!"),
         MISSING_BY("Include a timing you'll finish your task by! Using /by ..."),
         MISSING_BY_2("NonoNonoNO! Tell me WHEN you're gonna finish the task by!! Share with me the DATE!!"),
@@ -39,22 +41,22 @@ public class Ui {
         System.out.println(msg);
     }
 
-    void greet() {
+    public void greet() {
         showLine();
         System.out.println(KLEE_LOGO);
         System.out.println("Hello, I'm Klee, your bestest elf-friend! (◕‿◕✿) \nWhat can I do for you today? :3");
         showLine();
     }
 
-    void exit() {
+    public void exit() {
         System.out.println("Aww mann, you're leaving already?? (╥﹏╥) ...Well!! Come back soon to play with me again alright!!");
     }
 
-    void echo(String input) {
+    public void echo(String input) {
         System.out.println("\"" + input + "\"" + ", hehee!");
     }
 
-    void throwTantrum() {
+    public void throwTantrum() {
         System.out.println("Huh?? I'm not sure I quite understand o(╥﹏╥)o... could you be more specific please?");
     }
